Recast is_api_request_test as system test
